extend to monthly recurrence module
